requirements.txt
transformers
tokenizer
pytest
torch
tqdm